+++ RandomEncounters/Utilities/AssetLoader.cs
@@ -9,8 +9,7 @@ namespace RandomEncounters
     internal class AssetLoader
     {
         public static GameObject Hull { get; private set; }
-        public static GameObject Mast { get; private set; }
-        public static GameObject Bowsprit { get; private set; }
+        public static GameObject SmallWreck { get; private set; }
 
         internal static IEnumerator LoadAssetBundle()
         {
@@ -20,17 +19,16 @@ namespace RandomEncounters
             yield return assetBundleRequest;
 
             var assetBundle = assetBundleRequest.assetBundle;
-            if (assetBundle == null)            
+            if (assetBundle == null)
                 LogError($"Failed to load {bundlePath}");
             var request = assetBundle.LoadAllAssetsAsync();
             yield return request;
 
             Hull = request.allAssets.FirstOrDefault(a => a.name == "hull") as GameObject;
-            Mast = request.allAssets.FirstOrDefault(a => a.name == "mast") as GameObject;
-            Bowsprit = request.allAssets.FirstOrDefault(a => a.name == "bowsprit") as GameObject;
+            SmallWreck = request.allAssets.FirstOrDefault(a => a.name == "small_wreck") as GameObject;
 
-            if(Hull == null || Mast == null || Bowsprit == null)            
-                LogError("Failed to load all assets from the bundle.");            
+            if (Hull == null || SmallWreck == null)
+                LogError("Failed to load all assets from the bundle.");
         }
     }
 }